requirements.txt
ultralytics
fastapi
uvicorn[standard]
pillow
python-multipart